20.03.15) AbstractEntity 생성 및 중복 코드 이동
- User, Answer, Question extends AbstractEntity.
