+++ src/main/java/com/codessquad/qna/domain/Answer.java
@@ -1,22 +1,11 @@
 package com.codessquad.qna.domain;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import org.apache.commons.lang3.builder.ToStringBuilder;
-import org.apache.commons.lang3.builder.ToStringStyle;
 
 import javax.persistence.*;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 
 @Entity
-public class Answer {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @JsonProperty
-    private Long id;
-
+public class Answer extends AbstractEntity {
     @ManyToOne
     @JoinColumn(foreignKey = @ForeignKey(name = "fk_answer_writer"))
     @JsonProperty
@@ -25,8 +14,6 @@ public class Answer {
     @Lob
     @JsonProperty
     private String contents;
-    @JsonProperty
-    private LocalDateTime createdDate;
 
     @ManyToOne
     @JoinColumn(foreignKey = @ForeignKey(name = "fk_answer_question"))
@@ -42,7 +29,6 @@ public class Answer {
         this.question = question;
         this.contents = contents;
         this.writer = writer;
-        this.createdDate = LocalDateTime.now();
         this.deleted = false;
     }
 
@@ -52,14 +38,6 @@ public class Answer {
 
     public void setDeleted(boolean deleted) {
         this.deleted = deleted;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
     }
 
     public User getWriter() {
@@ -78,14 +56,6 @@ public class Answer {
         this.contents = contents;
     }
 
-    public LocalDateTime getCreatedDate() {
-        return createdDate;
-    }
-
-    public void setCreatedDate(LocalDateTime createdDate) {
-        this.createdDate = createdDate;
-    }
-
     public Question getQuestion() {
         return question;
     }
@@ -94,21 +64,8 @@ public class Answer {
         this.question = question;
     }
 
-    @Override
-    public String toString() {
-        return ToStringBuilder.reflectionToString(this, ToStringStyle.SIMPLE_STYLE);
-    }
-
     public void update(String contents) {
         this.contents = contents;
-        this.createdDate = LocalDateTime.now();
-    }
-
-    public String getFormattedCreatedDate() {
-        if (createdDate == null) {
-            return "";
-        }
-        return createdDate.format(DateTimeFormatter.ofPattern("YYYY-MM-dd HH:mm:ss"));
     }
 
     public void delete() {
