+++ src/main/resources/import.sql
@@ -1,5 +1,5 @@
-INSERT INTO USER (id, USER_ID, NAME, PASSWORD, EMAIL) VALUES (1, 'guswns1659', 'hyunjun', 'pass', 'guswns1659@gmail.com');
-INSERT INTO USER (id, USER_ID, NAME, PASSWORD, EMAIL) VALUES (2, 'zmdk1127', 'jun', 'pass', 'zmdk1127@naver.com');
+INSERT INTO USER (id, USER_ID, NAME, PASSWORD, EMAIL, CREATED_DATE) VALUES (1, 'guswns1659', 'hyunjun', 'pass', 'guswns1659@gmail.com', CURRENT_TIMESTAMP());
+INSERT INTO USER (id, USER_ID, NAME, PASSWORD, EMAIL, CREATED_DATE) VALUES (2, 'zmdk1127', 'jun', 'pass', 'zmdk1127@naver.com', CURRENT_TIMESTAMP());
 INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (1, 'CONTENTS1', CURRENT_TIMESTAMP(), FALSE ,'JAVA IS DIFFICULT', 1, 0);
 INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (2, 'CONTENTS2', CURRENT_TIMESTAMP(), FALSE ,'JAVA IS EASY', 2, 0);
