+++ src/main/resources/logback.xml
@@ -6,7 +6,7 @@
         </layout>
     </appender>
     <logger name="kr.pe.lahuman" level="TRACE"/>
-    <root level="info">
+    <root level="DEBUG">
         <appender-ref ref="STDOUT" />
     </root>
 </configuration>
